requirements.txt
opencv-python==4.10.0.84 
pyserial==3.5
requests==2.32.3
python-dotenv==0.19.1